pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.32"}
alembic = "^1.13.2"
psycopg2 = "^2.9.9"

[tool.poetry.dev-dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
